wgpu: Refactor out filters into their own files
