+++ render/wgpu/src/filters/color_matrix.rs
@@ -1,0 +1,186 @@
+use crate::backend::RenderTargetMode;
+use crate::buffer_pool::TexturePool;
+use crate::descriptors::Descriptors;
+use crate::filters::{create_filter_vertices, VERTEX_BUFFERS_DESCRIPTION_FILTERS};
+use crate::surface::target::CommandTarget;
+use crate::utils::SampleCountMap;
+use std::sync::OnceLock;
+use swf::ColorMatrixFilter as ColorMatrixFilterArgs;
+use wgpu::util::DeviceExt;
+
+pub struct ColorMatrixFilter {
+    bind_group_layout: wgpu::BindGroupLayout,
+    pipeline_layout: wgpu::PipelineLayout,
+    pipelines: SampleCountMap<OnceLock<wgpu::RenderPipeline>>,
+}
+
+impl ColorMatrixFilter {
+    pub fn new(device: &wgpu::Device) -> Self {
+        let bind_group_layout = device.create_bind_group_layout(&wgpu::BindGroupLayoutDescriptor {
+            entries: &[
+                wgpu::BindGroupLayoutEntry {
+                    binding: 0,
+                    visibility: wgpu::ShaderStages::FRAGMENT,
+                    ty: wgpu::BindingType::Texture {
+                        multisampled: false,
+                        sample_type: wgpu::TextureSampleType::Float { filterable: true },
+                        view_dimension: wgpu::TextureViewDimension::D2,
+                    },
+                    count: None,
+                },
+                wgpu::BindGroupLayoutEntry {
+                    binding: 1,
+                    visibility: wgpu::ShaderStages::FRAGMENT,
+                    ty: wgpu::BindingType::Sampler(wgpu::SamplerBindingType::Filtering),
+                    count: None,
+                },
+                wgpu::BindGroupLayoutEntry {
+                    binding: 2,
+                    visibility: wgpu::ShaderStages::FRAGMENT,
+                    ty: wgpu::BindingType::Buffer {
+                        ty: wgpu::BufferBindingType::Uniform,
+                        has_dynamic_offset: false,
+                        min_binding_size: wgpu::BufferSize::new(
+                            std::mem::size_of::<[f32; 20]>() as u64
+                        ),
+                    },
+                    count: None,
+                },
+            ],
+            label: create_debug_label!("Color matrix filter binds").as_deref(),
+        });
+
+        let pipeline_layout = device.create_pipeline_layout(&wgpu::PipelineLayoutDescriptor {
+            label: None,
+            bind_group_layouts: &[&bind_group_layout],
+            push_constant_ranges: &[],
+        });
+
+        Self {
+            pipelines: Default::default(),
+            pipeline_layout,
+            bind_group_layout,
+        }
+    }
+
+    fn pipeline(&self, descriptors: &Descriptors, msaa_sample_count: u32) -> &wgpu::RenderPipeline {
+        self.pipelines.get_or_init(msaa_sample_count, || {
+            let label = create_debug_label!("Color Matrix Filter ({} msaa)", msaa_sample_count);
+            descriptors
+                .device
+                .create_render_pipeline(&wgpu::RenderPipelineDescriptor {
+                    label: label.as_deref(),
+                    layout: Some(&self.pipeline_layout),
+                    vertex: wgpu::VertexState {
+                        module: &descriptors.shaders.color_matrix_filter,
+                        entry_point: "main_vertex",
+                        buffers: &VERTEX_BUFFERS_DESCRIPTION_FILTERS,
+                    },
+                    primitive: wgpu::PrimitiveState {
+                        topology: wgpu::PrimitiveTopology::TriangleList,
+                        strip_index_format: None,
+                        front_face: wgpu::FrontFace::Ccw,
+                        cull_mode: None,
+                        polygon_mode: wgpu::PolygonMode::default(),
+                        unclipped_depth: false,
+                        conservative: false,
+                    },
+                    depth_stencil: None,
+                    multisample: wgpu::MultisampleState {
+                        count: msaa_sample_count,
+                        mask: !0,
+                        alpha_to_coverage_enabled: false,
+                    },
+                    fragment: Some(wgpu::FragmentState {
+                        module: &descriptors.shaders.color_matrix_filter,
+                        entry_point: "main_fragment",
+                        targets: &[Some(wgpu::TextureFormat::Rgba8Unorm.into())],
+                    }),
+                    multiview: None,
+                })
+        })
+    }
+
+    #[allow(clippy::too_many_arguments)]
+    pub fn apply(
+        &self,
+        descriptors: &Descriptors,
+        texture_pool: &mut TexturePool,
+        draw_encoder: &mut wgpu::CommandEncoder,
+        source_texture: &wgpu::Texture,
+        source_point: (u32, u32),
+        source_size: (u32, u32),
+        filter: &ColorMatrixFilterArgs,
+    ) -> CommandTarget {
+        let sample_count = source_texture.sample_count();
+        let format = source_texture.format();
+        let pipeline = self.pipeline(descriptors, sample_count);
+
+        let target = CommandTarget::new(
+            descriptors,
+            texture_pool,
+            wgpu::Extent3d {
+                width: source_size.0,
+                height: source_size.1,
+                depth_or_array_layers: 1,
+            },
+            format,
+            sample_count,
+            RenderTargetMode::FreshWithColor(wgpu::Color::TRANSPARENT),
+            draw_encoder,
+        );
+        let source_view = source_texture.create_view(&Default::default());
+        let buffer = descriptors
+            .device
+            .create_buffer_init(&wgpu::util::BufferInitDescriptor {
+                label: create_debug_label!("Filter arguments").as_deref(),
+                contents: bytemuck::cast_slice(&filter.matrix),
+                usage: wgpu::BufferUsages::UNIFORM,
+            });
+        let vertices = create_filter_vertices(
+            &descriptors.device,
+            source_texture,
+            source_point,
+            source_size,
+        );
+        let filter_group = descriptors
+            .device
+            .create_bind_group(&wgpu::BindGroupDescriptor {
+                label: create_debug_label!("Filter group").as_deref(),
+                layout: &self.bind_group_layout,
+                entries: &[
+                    wgpu::BindGroupEntry {
+                        binding: 0,
+                        resource: wgpu::BindingResource::TextureView(&source_view),
+                    },
+                    wgpu::BindGroupEntry {
+                        binding: 1,
+                        resource: wgpu::BindingResource::Sampler(
+                            descriptors.bitmap_samplers.get_sampler(false, false),
+                        ),
+                    },
+                    wgpu::BindGroupEntry {
+                        binding: 2,
+                        resource: buffer.as_entire_binding(),
+                    },
+                ],
+            });
+        let mut render_pass = draw_encoder.begin_render_pass(&wgpu::RenderPassDescriptor {
+            label: create_debug_label!("Color matrix filter").as_deref(),
+            color_attachments: &[target.color_attachments()],
+            depth_stencil_attachment: None,
+        });
+        render_pass.set_pipeline(pipeline);
+
+        render_pass.set_bind_group(0, &filter_group, &[]);
+
+        render_pass.set_vertex_buffer(0, vertices.slice(..));
+        render_pass.set_index_buffer(
+            descriptors.quad.indices.slice(..),
+            wgpu::IndexFormat::Uint32,
+        );
+        render_pass.draw_indexed(0..6, 0, 0..1);
+        drop(render_pass);
+        target
+    }
+}
