+++ render/wgpu/src/surface.rs
@@ -7,17 +7,13 @@ use crate::buffer_pool::TexturePool;
 use crate::mesh::Mesh;
 use crate::surface::commands::{chunk_blends, Chunk, CommandRenderer};
 use crate::utils::{remove_srgb, supported_sample_count};
-use crate::{
-    ColorAdjustments, Descriptors, FilterVertex, MaskState, Pipelines, Transforms, UniformBuffer,
-};
+use crate::{ColorAdjustments, Descriptors, MaskState, Pipelines, Transforms, UniformBuffer};
 use ruffle_render::commands::CommandList;
 use ruffle_render::filters::Filter;
 use ruffle_render::quality::StageQuality;
 use std::sync::Arc;
-use swf::{BlurFilter, ColorMatrixFilter};
 use target::CommandTarget;
 use tracing::instrument;
-use wgpu::util::DeviceExt;
 
 use crate::utils::run_copy_pipeline;
 
@@ -351,7 +347,7 @@ impl Surface {
         filter: Filter,
     ) -> CommandTarget {
         let target = match filter {
-            Filter::ColorMatrixFilter(filter) => self.apply_color_matrix(
+            Filter::ColorMatrixFilter(filter) => descriptors.filters.color_matrix.apply(
                 descriptors,
                 texture_pool,
                 draw_encoder,
@@ -360,7 +356,7 @@ impl Surface {
                 source_size,
                 &filter,
             ),
-            Filter::BlurFilter(filter) => self.apply_blur(
+            Filter::BlurFilter(filter) => descriptors.filters.blur.apply(
                 descriptors,
                 texture_pool,
                 draw_encoder,
@@ -372,7 +368,7 @@ impl Surface {
             _ => {
                 tracing::warn!("Unsupported filter {filter:?}");
                 // Apply a default color matrix - it's essentially a blit
-                self.apply_color_matrix(
+                descriptors.filters.color_matrix.apply(
                     descriptors,
                     texture_pool,
                     draw_encoder,
@@ -390,246 +386,4 @@ impl Surface {
         target.ensure_cleared(draw_encoder);
         target
     }
-
-    #[allow(clippy::too_many_arguments)]
-    pub fn apply_color_matrix(
-        &self,
-        descriptors: &Descriptors,
-        texture_pool: &mut TexturePool,
-        draw_encoder: &mut wgpu::CommandEncoder,
-        source_texture: &wgpu::Texture,
-        source_point: (u32, u32),
-        source_size: (u32, u32),
-        filter: &ColorMatrixFilter,
-    ) -> CommandTarget {
-        let target = CommandTarget::new(
-            descriptors,
-            texture_pool,
-            wgpu::Extent3d {
-                width: source_size.0,
-                height: source_size.1,
-                depth_or_array_layers: 1,
-            },
-            self.format,
-            self.sample_count,
-            RenderTargetMode::FreshWithColor(wgpu::Color::TRANSPARENT),
-            draw_encoder,
-        );
-        let source_view = source_texture.create_view(&Default::default());
-        let buffer = descriptors
-            .device
-            .create_buffer_init(&wgpu::util::BufferInitDescriptor {
-                label: create_debug_label!("Filter arguments").as_deref(),
-                contents: bytemuck::cast_slice(&filter.matrix),
-                usage: wgpu::BufferUsages::UNIFORM,
-            });
-        let vertices = create_filter_vertices(
-            &descriptors.device,
-            source_texture,
-            source_point,
-            source_size,
-        );
-        let filter_group = descriptors
-            .device
-            .create_bind_group(&wgpu::BindGroupDescriptor {
-                label: create_debug_label!("Filter group").as_deref(),
-                layout: &descriptors.bind_layouts.color_matrix_filter,
-                entries: &[
-                    wgpu::BindGroupEntry {
-                        binding: 0,
-                        resource: wgpu::BindingResource::TextureView(&source_view),
-                    },
-                    wgpu::BindGroupEntry {
-                        binding: 1,
-                        resource: wgpu::BindingResource::Sampler(
-                            descriptors.bitmap_samplers.get_sampler(false, false),
-                        ),
-                    },
-                    wgpu::BindGroupEntry {
-                        binding: 2,
-                        resource: buffer.as_entire_binding(),
-                    },
-                ],
-            });
-        let mut render_pass = draw_encoder.begin_render_pass(&wgpu::RenderPassDescriptor {
-            label: create_debug_label!("Color matrix filter").as_deref(),
-            color_attachments: &[target.color_attachments()],
-            depth_stencil_attachment: None,
-        });
-        render_pass.set_pipeline(&self.pipelines.color_matrix_filter);
-
-        render_pass.set_bind_group(0, &filter_group, &[]);
-
-        render_pass.set_vertex_buffer(0, vertices.slice(..));
-        render_pass.set_index_buffer(
-            descriptors.quad.indices.slice(..),
-            wgpu::IndexFormat::Uint32,
-        );
-        render_pass.draw_indexed(0..6, 0, 0..1);
-        drop(render_pass);
-        target
-    }
-
-    #[allow(clippy::too_many_arguments)]
-    pub fn apply_blur(
-        &self,
-        descriptors: &Descriptors,
-        texture_pool: &mut TexturePool,
-        draw_encoder: &mut wgpu::CommandEncoder,
-        source_texture: &wgpu::Texture,
-        source_point: (u32, u32),
-        source_size: (u32, u32),
-        filter: &BlurFilter,
-    ) -> CommandTarget {
-        // FIXME - this should be larger than the source texture. Figure out exactly how much larger
-        let targets = [
-            CommandTarget::new(
-                descriptors,
-                texture_pool,
-                wgpu::Extent3d {
-                    width: source_size.0,
-                    height: source_size.1,
-                    depth_or_array_layers: 1,
-                },
-                self.format,
-                self.sample_count,
-                RenderTargetMode::FreshWithColor(wgpu::Color::TRANSPARENT),
-                draw_encoder,
-            ),
-            CommandTarget::new(
-                descriptors,
-                texture_pool,
-                wgpu::Extent3d {
-                    width: source_size.0,
-                    height: source_size.1,
-                    depth_or_array_layers: 1,
-                },
-                self.format,
-                self.sample_count,
-                RenderTargetMode::FreshWithColor(wgpu::Color::TRANSPARENT),
-                draw_encoder,
-            ),
-        ];
-
-        // TODO: Vertices should be per pass, and each pass needs diff sizes
-        let vertices = create_filter_vertices(
-            &descriptors.device,
-            source_texture,
-            source_point,
-            source_size,
-        );
-
-        let source_view = source_texture.create_view(&Default::default());
-        for i in 0..2 {
-            let blur_x = (filter.blur_x.to_f32() - 1.0).max(0.0);
-            let blur_y = (filter.blur_y.to_f32() - 1.0).max(0.0);
-            let current = &targets[i % 2];
-            let (previous_view, previous_vertices, previous_width, previous_height) = if i == 0 {
-                (
-                    &source_view,
-                    vertices.slice(..),
-                    source_texture.width() as f32,
-                    source_texture.height() as f32,
-                )
-            } else {
-                let previous = &targets[(i - 1) % 2];
-                (
-                    previous.color_view(),
-                    descriptors.quad.filter_vertices.slice(..),
-                    previous.width() as f32,
-                    previous.height() as f32,
-                )
-            };
-            let buffer = descriptors
-                .device
-                .create_buffer_init(&wgpu::util::BufferInitDescriptor {
-                    label: create_debug_label!("Filter arguments").as_deref(),
-                    contents: bytemuck::cast_slice(&[
-                        blur_x * ((i as u32) % 2) as f32,
-                        blur_y * (((i as u32) % 2) + 1) as f32,
-                        previous_width,
-                        previous_height,
-                    ]),
-                    usage: wgpu::BufferUsages::UNIFORM,
-                });
-            let filter_group = descriptors
-                .device
-                .create_bind_group(&wgpu::BindGroupDescriptor {
-                    label: create_debug_label!("Filter group").as_deref(),
-                    layout: &descriptors.bind_layouts.blur_filter,
-                    entries: &[
-                        wgpu::BindGroupEntry {
-                            binding: 0,
-                            resource: wgpu::BindingResource::TextureView(previous_view),
-                        },
-                        wgpu::BindGroupEntry {
-                            binding: 1,
-                            resource: wgpu::BindingResource::Sampler(
-                                descriptors.bitmap_samplers.get_sampler(false, true),
-                            ),
-                        },
-                        wgpu::BindGroupEntry {
-                            binding: 2,
-                            resource: buffer.as_entire_binding(),
-                        },
-                    ],
-                });
-            let mut render_pass = draw_encoder.begin_render_pass(&wgpu::RenderPassDescriptor {
-                label: create_debug_label!("Blur filter").as_deref(),
-                color_attachments: &[current.color_attachments()],
-                depth_stencil_attachment: None,
-            });
-            render_pass.set_pipeline(&self.pipelines.blur_filter);
-
-            render_pass.set_bind_group(0, &filter_group, &[]);
-
-            render_pass.set_vertex_buffer(0, previous_vertices);
-            render_pass.set_index_buffer(
-                descriptors.quad.indices.slice(..),
-                wgpu::IndexFormat::Uint32,
-            );
-            render_pass.draw_indexed(0..6, 0, 0..1);
-        }
-
-        targets
-            .into_iter()
-            .last()
-            .expect("Targets should not be empty")
-    }
-}
-
-fn create_filter_vertices(
-    device: &wgpu::Device,
-    source_texture: &wgpu::Texture,
-    source_point: (u32, u32),
-    source_size: (u32, u32),
-) -> wgpu::Buffer {
-    let source_width = source_texture.width() as f32;
-    let source_height = source_texture.height() as f32;
-    let left = source_point.0;
-    let top = source_point.1;
-    let right = left + source_size.0;
-    let bottom = top + source_size.1;
-    device.create_buffer_init(&wgpu::util::BufferInitDescriptor {
-        label: create_debug_label!("Filter vertices").as_deref(),
-        contents: bytemuck::cast_slice(&[
-            FilterVertex {
-                position: [0.0, 0.0],
-                uv: [left as f32 / source_width, top as f32 / source_height],
-            },
-            FilterVertex {
-                position: [1.0, 0.0],
-                uv: [right as f32 / source_width, top as f32 / source_height],
-            },
-            FilterVertex {
-                position: [1.0, 1.0],
-                uv: [right as f32 / source_width, bottom as f32 / source_height],
-            },
-            FilterVertex {
-                position: [0.0, 1.0],
-                uv: [left as f32 / source_width, bottom as f32 / source_height],
-            },
-        ]),
-        usage: wgpu::BufferUsages::VERTEX,
-    })
 }
